requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.1.0
itsdangerous==2.1.2
Werkzeug==3.0.1
psycopg2-binary==2.9.9







